Cleanup updated

+++ cutile-compiler/src/compiler/compile_intrinsic.rs
@@ -129,6 +129,7 @@ impl<'m> CUDATileFunctionCompiler<'m> {
         })
     }
 
+    #[allow(clippy::too_many_arguments)]
     fn compile_nested_mutable_access_offset_metadata(
         &self,
         module: &mut Module,
@@ -194,7 +195,7 @@ impl<'m> CUDATileFunctionCompiler<'m> {
                 pid
             } else {
                 let ratio_value =
-                    self.compile_constant(module, block_id, generic_vars, ratio as i32)?;
+                    self.compile_constant(module, block_id, generic_vars, ratio)?;
                 self.compile_binary_op_from_values(
                     module,
                     block_id,
